requirements.txt
duckdb		
pandas		
requests		
jinja2		
pillow
psycopg2-binary
python-dateutil